some small refactoring and changes
