--- src/TimsShop/Controllers/StorageSettable.java
@@ -1,18 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package TimsShop.Controllers;
-
-import TimsShop.Models.DataModels.ShopDataStorage;
-
-/**
- *
- * @author win
- */
-public interface StorageSettable {
-    
-    public  void setStorage(ShopDataStorage storage);
-    
-}
