+++ src/TimsShop/FXML/SupplierView.fxml
@@ -11,7 +11,7 @@
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane id="AnchorPane" prefHeight="529.0" prefWidth="539.0" style="-fx-background-color: #ddd2d2;" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="TimsShop.Controllers.Dialogs.SupplierControllers.SupplierView">
+<AnchorPane id="AnchorPane" prefHeight="529.0" prefWidth="539.0" style="-fx-background-color: #ddd2d2;" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="TimsShop.Views.SupplierViews.SupplierView">
    <children>
       <GridPane layoutX="14.0" layoutY="14.0" prefHeight="505.0" prefWidth="564.0">
          <columnConstraints>
