+++ src/TimsShop/Views/SupplierViews/SupplierView.java
@@ -1,5 +1,5 @@
 
-package TimsShop.Controllers.Dialogs.SupplierControllers;
+package TimsShop.Views.SupplierViews;
 
 import TimsShop.Controllers.CallBackEvt;
 import TimsShop.Controllers.ViewLoader;
